producer/requirements.txt
aiobotocore==2.21.1
alembic==1.15.1
APScheduler==3.11.0
asyncpg==0.30.0
fastapi==0.115.12
kafka-python==2.1.2
pandas==2.2.3
pydantic==2.11.1
pydantic-settings==2.0.0
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
slowapi==0.1.9
SQLAlchemy==2.0.39
uvicorn==0.34.0
PyYaml==6.0.2